test: Rename and consolidate tests
